🎨 Style: 추후 구조 리팩토링을 위한 루트 레이아웃 분리
